Use envtest for tablet node component tests
Reduce pile of s-t.

Replace the brittle fake testing setup for tablet node tests with a real
controller-runtime envtest control plane so tests exercise production
code paths and Kubernetes-backed behavior.

Eliminate the FakeServer test shim and the fake client usage to validate
the real server implementation and pod listing/replica logic.

Add a shared envtest test environment in pkg/components/suite_test.go that
starts a control plane, installs CRDs, registers schemes,
and provides a real client.Client and rest.Config for component tests.

Remove the FakeServer test implementation and stop using
controller-runtime’s fake client in tablet-node tests.

Update pkg/components/tablet_node_test.go to create an isolated namespace
per test, create the Ytsaurus resource via the envtest API server,
and use apiproxy.NewYtsaurus with the real test client and scheme.

Add provisionTabletNode helper that calls the production server.Fetch
and server.Sync to provision the tablet node server state during tests,
and increase the recorder buffer to avoid blocked events.
