+++ pkg/components/suite_test.go
@@ -1,244 +1,21 @@
 package components
 
 import (
-	"context"
 	"os"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"go.ytsaurus.tech/yt/go/yt"
-
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-
-	ytv1 "github.com/ytsaurus/ytsaurus-k8s-operator/api/v1"
-
-	"github.com/ytsaurus/ytsaurus-k8s-operator/pkg/consts"
-	"github.com/ytsaurus/ytsaurus-k8s-operator/pkg/labeller"
-	mock_yt "github.com/ytsaurus/ytsaurus-k8s-operator/pkg/mock"
-	"github.com/ytsaurus/ytsaurus-k8s-operator/pkg/ypatch"
 )
 
 func TestComponents(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Components fake suite")
+	RunSpecs(t, "Components suite")
 }
 
 var _ = BeforeSuite(func() {
 	logf.SetLogger(zap.New(zap.WriteTo(os.Stderr), zap.UseDevMode(true)))
 })
-
-type FakeComponent struct {
-	name     string
-	compType consts.ComponentType
-	status   ComponentStatus
-}
-
-func NewFakeComponent(name string, compType consts.ComponentType) *FakeComponent {
-	return &FakeComponent{
-		name:     name,
-		compType: compType,
-		status:   ComponentStatusReady(),
-	}
-}
-
-func (fc *FakeComponent) Fetch(ctx context.Context) error {
-	return nil
-}
-
-func (fc *FakeComponent) Exists() bool {
-	return true
-}
-
-func (fc *FakeComponent) ArePodsReady(ctx context.Context) (ComponentStatus, error) {
-	return ComponentStatusReady(), nil
-}
-
-func (fc *FakeComponent) Sync(ctx context.Context, dry bool) (ComponentStatus, error) {
-	return fc.status, nil
-}
-
-func (fc *FakeComponent) GetStatus() ComponentStatus {
-	return fc.status
-}
-
-func (fc *FakeComponent) SetStatus(status ComponentStatus) {
-	fc.status = status
-}
-
-func (fc *FakeComponent) NeedSync() bool {
-	return false
-}
-
-func (fc *FakeComponent) NeedUpdate() ComponentStatus {
-	return ComponentStatusReady()
-}
-
-func (fc *FakeComponent) IsUpdating() bool {
-	return false
-}
-
-func (fc *FakeComponent) GetShortName() string {
-	return fc.name
-}
-
-func (fc *FakeComponent) GetFullName() string {
-	return fc.name
-}
-
-func (fc *FakeComponent) GetType() consts.ComponentType {
-	return fc.compType
-}
-
-func (fc *FakeComponent) GetComponent() ytv1.Component {
-	return ytv1.Component{
-		Type: fc.compType,
-		Name: fc.name,
-	}
-}
-
-func (fc *FakeComponent) GetLabeller() *labeller.Labeller {
-	return nil
-}
-
-func (fc *FakeComponent) GetCypressPatch() ypatch.PatchSet {
-	return nil
-}
-
-func (fc *FakeComponent) GetImageHeaterTarget() *ImageHeaterTarget {
-	return nil
-}
-
-func (fc *FakeComponent) GetReadyCondition() ComponentStatus {
-	return fc.status
-}
-
-func (fc *FakeComponent) SetReadyCondition(status ComponentStatus) {}
-
-type FakeServer struct {
-	instanceCount int32
-}
-
-var _ server = (*FakeServer)(nil)
-
-func NewFakeServer() *FakeServer {
-	return &FakeServer{}
-}
-
-func (fs *FakeServer) Fetch(ctx context.Context) error {
-	return nil
-}
-
-func (fs *FakeServer) needUpdate() ComponentStatus {
-	return ComponentStatusReady()
-}
-
-func (fs *FakeServer) getPodAnnotation(name string) string {
-	return ""
-}
-
-func (fs *FakeServer) podsImageCorrespondsToSpec() bool {
-	return true
-}
-
-func (fs *FakeServer) Exists() bool {
-	return true
-}
-
-func (fs *FakeServer) needSync(updating bool) bool {
-	return false
-}
-
-func (fs *FakeServer) getImageHeaterTarget() *ImageHeaterTarget {
-	return nil
-}
-
-func (fs *FakeServer) arePodsUpdatedToNewRevision(ctx context.Context) bool {
-	return true
-}
-
-func (fs *FakeServer) setUpdateStrategy(strategy appsv1.StatefulSetUpdateStrategyType, partition, maxUnavailable int32) {
-	// No-op for fake server:
-}
-
-func (fs *FakeServer) getInstanceCount() int32 {
-	return fs.instanceCount
-}
-
-func (fs *FakeServer) setInstanceCount(instanceCount int32) {
-	fs.instanceCount = instanceCount
-}
-
-func (fs *FakeServer) getMinReadyInstanceCount(margin int32) int32 {
-	return fs.instanceCount
-}
-
-func (fs *FakeServer) listPods(ctx context.Context) ([]corev1.Pod, error) {
-	return nil, nil
-}
-
-func (fs *FakeServer) getRollingUpdateStatus(ctx context.Context) (*stsRollingStatus, error) {
-	return nil, nil
-}
-
-func (fs *FakeServer) Sync(ctx context.Context) error {
-	return nil
-}
-
-func (fs *FakeServer) buildStatefulSet() *appsv1.StatefulSet {
-	return nil
-}
-
-func (fs *FakeServer) rebuildStatefulSet() *appsv1.StatefulSet {
-	return nil
-}
-
-func (fs *FakeServer) removePods(ctx context.Context) error {
-	return nil
-}
-
-func (fs *FakeServer) addCARootBundle(c *corev1.Container) {
-}
-
-func (fs *FakeServer) addTlsSecretMount(c *corev1.Container) {
-}
-
-func (fs *FakeServer) addMonitoringPort(port corev1.ServicePort) {
-}
-
-func (fs *FakeServer) GetImage() string {
-	return ""
-}
-
-type FakeYtsaurusClient struct {
-	FakeComponent
-	client *mock_yt.MockClient
-}
-
-func NewFakeYtsaurusClient(client *mock_yt.MockClient) *FakeYtsaurusClient {
-	return &FakeYtsaurusClient{
-		FakeComponent: *NewFakeComponent("ytsaurus_client", consts.YtsaurusClientType),
-		client:        client,
-	}
-}
-
-func (fyc *FakeYtsaurusClient) GetYtClient() yt.Client {
-	return fyc.client
-}
-
-func (fyc *FakeYtsaurusClient) SetStatus(status ComponentStatus) {
-	fyc.status = status
-}
-
-func (fyc *FakeYtsaurusClient) UpdatePreCheck(ctx context.Context) ComponentStatus {
-	return ComponentStatusReady()
-}
-
-func (fyc *FakeYtsaurusClient) ShouldSkipCypressOperations() bool {
-	return false
-}
